+++ lib/core/decision_making_helper.dart
@@ -1,3 +1,5 @@
+import 'dart:developer' as dev;
+
 import 'package:uuid/uuid.dart';
 
 /// AHP HELPER
@@ -5,5 +7,12 @@ class DecisionMakingHelper {
   /// HELPER TO GET UNIQUE ID
   String getCustomUniqueId() {
     return Uuid().v4();
+  }
+
+  void printLog({
+    required String message,
+    String? logName,
+  }) {
+    dev.log(message, name: logName ?? 'DECISION MAKING');
   }
 }