+++ lib/core/isolate/decision_isolate_message.dart
@@ -2,13 +2,13 @@ import 'dart:isolate';
 
 import 'package:flutter_decision_making/core/decision_making_enums.dart';
 
-class DecisionIsolatedMessage {
+class DecisionIsolateMessage {
   final DecisionAlgorithm algorithm;
   final dynamic command;
   final Map<String, dynamic> payload;
   final SendPort replyPort;
 
-  DecisionIsolatedMessage({
+  DecisionIsolateMessage({
     required this.algorithm,
     required this.command,
     required this.payload,